Moved doLevelUp to BaseClass, generified it to work with all subclasses.
